requirements.txt
flask>=3.1
fastapi>=0.115
uvicorn[standard]>=0.23
python-multipart>=0.0.6
jinja2>=3.1
weasyprint>=62.3
requests>=2.32
pandas>=2.3
openpyxl>=3.1
pydantic-settings>=2.6
